+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to VibeBar will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- unify update channel to use single appcast.xml for all users
+- stable releases now also update appcast-beta.xml to allow beta users to upgrade
+
 ## [1.3.5] - 2026-03-17
 
 ### Added